change structure and css
